requirements.txt
django==2.0.6
djangorestframework==3.8.2
django-filter==1.1.0
django-oauth-toolkit==1.2.0
djangorestframework-jsonapi==2.4.0
markdown==2.6.11
pytz==2018.4